requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.1.2
Django==5.2.10
django-admin-rangefilter==0.13.5
django-allauth==65.13.1
django-allauth-ui==1.8.1
django-tenants==3.9.0
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reportlab==4.4.10
six==1.17.0
slippers==0.6.2
sqlparse==0.5.5
typeguard==2.13.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
